feat: tabs layout 모바일 대응
